Fix Indices Request Cache ITs from flaking and timing out (#14369)
(cherry picked from commit bd56456baab283570da9b423152125124d482db3)
