+++ server/src/internalClusterTest/java/org/opensearch/indices/IndicesRequestCacheIT.java
@@ -37,7 +37,6 @@ import com.carrotsearch.randomizedtesting.annotations.ParametersFactory;
 import org.opensearch.action.admin.cluster.health.ClusterHealthResponse;
 import org.opensearch.action.admin.cluster.node.stats.NodeStats;
 import org.opensearch.action.admin.cluster.node.stats.NodesStatsResponse;
-import org.opensearch.action.admin.cluster.settings.ClusterUpdateSettingsRequest;
 import org.opensearch.action.admin.indices.alias.Alias;
 import org.opensearch.action.admin.indices.cache.clear.ClearIndicesCacheRequest;
 import org.opensearch.action.admin.indices.forcemerge.ForceMergeResponse;
@@ -55,7 +54,7 @@ import org.opensearch.common.util.FeatureFlags;
 import org.opensearch.core.index.Index;
 import org.opensearch.core.index.shard.ShardId;
 import org.opensearch.env.NodeEnvironment;
-import org.opensearch.index.IndexNotFoundException;
+import org.opensearch.index.IndexSettings;
 import org.opensearch.index.cache.request.RequestCacheStats;
 import org.opensearch.index.query.QueryBuilders;
 import org.opensearch.search.aggregations.bucket.global.GlobalAggregationBuilder;
@@ -75,13 +74,10 @@ import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_NUMBER_OF_REPLICAS;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_NUMBER_OF_SHARDS;
 import static org.opensearch.cluster.routing.allocation.decider.EnableAllocationDecider.CLUSTER_ROUTING_ALLOCATION_ENABLE_SETTING;
-import static org.opensearch.indices.IndicesRequestCache.INDICES_REQUEST_CACHE_STALENESS_THRESHOLD_SETTING;
-import static org.opensearch.indices.IndicesService.INDICES_CACHE_CLEANUP_INTERVAL_SETTING_KEY;
 import static org.opensearch.search.SearchService.CLUSTER_CONCURRENT_SEGMENT_SEARCH_SETTING;
 import static org.opensearch.search.aggregations.AggregationBuilders.dateHistogram;
 import static org.opensearch.search.aggregations.AggregationBuilders.dateRange;
@@ -126,6 +122,8 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
                         .put(IndicesRequestCache.INDEX_CACHE_REQUEST_ENABLED_SETTING.getKey(), true)
                         .put(SETTING_NUMBER_OF_SHARDS, 1)
                         .put(SETTING_NUMBER_OF_REPLICAS, 0)
+                        // Disable index refreshing to avoid cache being invalidated mid-test
+                        .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), TimeValue.timeValueMillis(-1))
                 )
                 .get()
         );
@@ -135,6 +133,8 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
             client.prepareIndex(index).setSource("f", "2014-05-13T00:00:00.000Z")
         );
         ensureSearchable(index);
+        // Force merge the index to ensure there can be no background merges during the subsequent searches that would invalidate the cache
+        forceMerge(client, index);
 
         // This is not a random example: serialization with time zones writes shared strings
         // which used to not work well with the query cache because of the handles stream output
@@ -197,6 +197,8 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
                         .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 5)
                         .put("index.number_of_routing_shards", 5)
                         .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
+                        // Disable index refreshing to avoid cache being invalidated mid-test
+                        .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), TimeValue.timeValueMillis(-1))
                 )
                 .get()
         );
@@ -214,12 +216,8 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
         );
         ensureSearchable(index);
         assertCacheState(client, index, 0, 0);
-
         // Force merge the index to ensure there can be no background merges during the subsequent searches that would invalidate the cache
-        ForceMergeResponse forceMergeResponse = client.admin().indices().prepareForceMerge(index).setFlush(true).get();
-        OpenSearchAssertions.assertAllSuccessful(forceMergeResponse);
-        refreshAndWaitForReplication();
-        ensureSearchable(index);
+        forceMerge(client, index);
 
         assertCacheState(client, index, 0, 0);
 
@@ -268,6 +266,8 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
                         .put(IndicesRequestCache.INDEX_CACHE_REQUEST_ENABLED_SETTING.getKey(), true)
                         .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 1)
                         .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
+                        // Disable index refreshing to avoid cache being invalidated mid-test
+                        .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), TimeValue.timeValueMillis(-1))
                 )
                 .get()
         );
@@ -287,10 +287,7 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
         assertCacheState(client, index, 0, 0);
 
         // Force merge the index to ensure there can be no background merges during the subsequent searches that would invalidate the cache
-        ForceMergeResponse forceMergeResponse = client.admin().indices().prepareForceMerge(index).setFlush(true).get();
-        OpenSearchAssertions.assertAllSuccessful(forceMergeResponse);
-        refreshAndWaitForReplication();
-        ensureSearchable(index);
+        forceMerge(client, index);
 
         assertCacheState(client, index, 0, 0);
 
@@ -335,6 +332,8 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
                         .put(IndicesRequestCache.INDEX_CACHE_REQUEST_ENABLED_SETTING.getKey(), true)
                         .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 1)
                         .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
+                        // Disable index refreshing to avoid cache being invalidated mid-test
+                        .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), TimeValue.timeValueMillis(-1))
                 )
                 .get()
         );
@@ -354,10 +353,7 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
         assertCacheState(client, index, 0, 0);
 
         // Force merge the index to ensure there can be no background merges during the subsequent searches that would invalidate the cache
-        ForceMergeResponse forceMergeResponse = client.admin().indices().prepareForceMerge(index).setFlush(true).get();
-        OpenSearchAssertions.assertAllSuccessful(forceMergeResponse);
-        refreshAndWaitForReplication();
-        ensureSearchable(index);
+        forceMerge(client, index);
 
         assertCacheState(client, index, 0, 0);
 
@@ -399,6 +395,8 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
             .put(IndicesRequestCache.INDEX_CACHE_REQUEST_ENABLED_SETTING.getKey(), true)
             .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 1)
             .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
+            // Disable index refreshing to avoid cache being invalidated mid-test
+            .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), TimeValue.timeValueMillis(-1))
             .build();
         assertAcked(client.admin().indices().prepareCreate("index-1").setMapping("d", "type=date").setSettings(settings).get());
         assertAcked(client.admin().indices().prepareCreate("index-2").setMapping("d", "type=date").setSettings(settings).get());
@@ -480,6 +478,7 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
             .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 2)
             .put("index.number_of_routing_shards", 2)
             .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
+            .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), TimeValue.timeValueMillis(-1))
             .build();
         String index = "index";
         assertAcked(client.admin().indices().prepareCreate(index).setMapping("s", "type=date").setSettings(settings).get());
@@ -499,10 +498,7 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
         assertCacheState(client, index, 0, 0);
 
         // Force merge the index to ensure there can be no background merges during the subsequent searches that would invalidate the cache
-        ForceMergeResponse forceMergeResponse = client.admin().indices().prepareForceMerge(index).setFlush(true).get();
-        OpenSearchAssertions.assertAllSuccessful(forceMergeResponse);
-        refreshAndWaitForReplication();
-        ensureSearchable(index);
+        forceMerge(client, index);
 
         assertCacheState(client, index, 0, 0);
 
@@ -644,11 +640,15 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
                         .put(IndicesRequestCache.INDEX_CACHE_REQUEST_ENABLED_SETTING.getKey(), true)
                         .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 1)
                         .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
+                        // Disable index refreshing to avoid cache being invalidated mid-test
+                        .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), TimeValue.timeValueMillis(-1))
                 )
                 .get()
         );
         indexRandom(true, client.prepareIndex(index).setSource("k", "hello"));
         ensureSearchable(index);
+        // Force merge the index to ensure there can be no background merges during the subsequent searches that would invalidate the cache
+        forceMerge(client, index);
 
         int expectedHits = 0;
         int expectedMisses = 0;
@@ -673,555 +673,6 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
         }
     }
 
-    public void testCacheWithInvalidation() throws Exception {
-        Client client = client();
-        String index = "index";
-        assertAcked(
-            client.admin()
-                .indices()
-                .prepareCreate(index)
-                .setMapping("k", "type=keyword")
-                .setSettings(
-                    Settings.builder()
-                        .put(IndicesRequestCache.INDEX_CACHE_REQUEST_ENABLED_SETTING.getKey(), true)
-                        .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 1)
-                        .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
-                )
-                .get()
-        );
-        indexRandom(true, client.prepareIndex(index).setSource("k", "hello"));
-        ensureSearchable(index);
-        SearchResponse resp = client.prepareSearch(index).setRequestCache(true).setQuery(QueryBuilders.termQuery("k", "hello")).get();
-        assertSearchResponse(resp);
-        OpenSearchAssertions.assertAllSuccessful(resp);
-        assertThat(resp.getHits().getTotalHits().value, equalTo(1L));
-
-        assertCacheState(client, index, 0, 1);
-        // Index but don't refresh
-        indexRandom(false, client.prepareIndex(index).setSource("k", "hello2"));
-        resp = client.prepareSearch(index).setRequestCache(true).setQuery(QueryBuilders.termQuery("k", "hello")).get();
-        assertSearchResponse(resp);
-        // Should expect hit as here as refresh didn't happen
-        assertCacheState(client, index, 1, 1);
-
-        // Explicit refresh would invalidate cache
-        refreshAndWaitForReplication();
-        // Hit same query again
-        resp = client.prepareSearch(index).setRequestCache(true).setQuery(QueryBuilders.termQuery("k", "hello")).get();
-        assertSearchResponse(resp);
-        // Should expect miss as key has changed due to change in IndexReader.CacheKey (due to refresh)
-        assertCacheState(client, index, 1, 2);
-    }
-
-    // calling cache clear api, when staleness threshold is lower than staleness, it should clean the stale keys from cache
-    public void testCacheClearAPIRemovesStaleKeysWhenStalenessThresholdIsLow() throws Exception {
-        String node = internalCluster().startNode(
-            Settings.builder()
-                .put(IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_STALENESS_THRESHOLD_SETTING_KEY, 0.10)
-                .put(
-                    IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_INTERVAL_SETTING_KEY,
-                    // setting intentionally high to avoid cache cleaner interfering
-                    TimeValue.timeValueMillis(300)
-                )
-        );
-        Client client = client(node);
-        String index1 = "index1";
-        String index2 = "index2";
-        setupIndex(client, index1);
-        setupIndex(client, index2);
-
-        // create first cache entry in index1
-        createCacheEntry(client, index1, "hello");
-        assertCacheState(client, index1, 0, 1);
-        long memorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(memorySizeForIndex1 > 0);
-
-        // create second cache entry in index1
-        createCacheEntry(client, index1, "there");
-        assertCacheState(client, index1, 0, 2);
-        long finalMemorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(finalMemorySizeForIndex1 > memorySizeForIndex1);
-
-        // create first cache entry in index2
-        createCacheEntry(client, index2, "hello");
-        assertCacheState(client, index2, 0, 1);
-        assertTrue(getRequestCacheStats(client, index2).getMemorySizeInBytes() > 0);
-
-        ClearIndicesCacheRequest clearIndicesCacheRequest = new ClearIndicesCacheRequest(index2);
-        client.admin().indices().clearCache(clearIndicesCacheRequest).actionGet();
-
-        // cache cleaner should have cleaned up the stale key from index 2
-        assertEquals(0, getRequestCacheStats(client, index2).getMemorySizeInBytes());
-        // cache cleaner should NOT have cleaned from index 1
-        assertEquals(finalMemorySizeForIndex1, getRequestCacheStats(client, index1).getMemorySizeInBytes());
-    }
-
-    // when staleness threshold is lower than staleness, it should clean the stale keys from cache
-    public void testStaleKeysCleanupWithLowThreshold() throws Exception {
-        int cacheCleanIntervalInMillis = 1;
-        String node = internalCluster().startNode(
-            Settings.builder()
-                .put(IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_STALENESS_THRESHOLD_SETTING_KEY, 0.10)
-                .put(
-                    IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_INTERVAL_SETTING_KEY,
-                    TimeValue.timeValueMillis(cacheCleanIntervalInMillis)
-                )
-        );
-        Client client = client(node);
-        String index1 = "index1";
-        String index2 = "index2";
-        setupIndex(client, index1);
-        setupIndex(client, index2);
-
-        // create first cache entry in index1
-        createCacheEntry(client, index1, "hello");
-        assertCacheState(client, index1, 0, 1);
-        long memorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(memorySizeForIndex1 > 0);
-
-        // create second cache entry in index1
-        createCacheEntry(client, index1, "there");
-        assertCacheState(client, index1, 0, 2);
-        long finalMemorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(finalMemorySizeForIndex1 > memorySizeForIndex1);
-
-        // create first cache entry in index2
-        createCacheEntry(client, index2, "hello");
-        assertCacheState(client, index2, 0, 1);
-        assertTrue(getRequestCacheStats(client, index2).getMemorySizeInBytes() > 0);
-
-        // force refresh so that it creates 1 stale key
-        flushAndRefresh(index2);
-        // sleep until cache cleaner would have cleaned up the stale key from index 2
-        assertBusy(() -> {
-            // cache cleaner should have cleaned up the stale key from index 2
-            assertEquals(0, getRequestCacheStats(client, index2).getMemorySizeInBytes());
-            // cache cleaner should NOT have cleaned from index 1
-            assertEquals(finalMemorySizeForIndex1, getRequestCacheStats(client, index1).getMemorySizeInBytes());
-        }, cacheCleanIntervalInMillis * 2, TimeUnit.MILLISECONDS);
-        // sleep until cache cleaner would have cleaned up the stale key from index 2
-    }
-
-    // when staleness threshold is equal to staleness, it should clean the stale keys from cache
-    public void testCacheCleanupOnEqualStalenessAndThreshold() throws Exception {
-        int cacheCleanIntervalInMillis = 1;
-        String node = internalCluster().startNode(
-            Settings.builder()
-                .put(IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_STALENESS_THRESHOLD_SETTING_KEY, 0.33)
-                .put(
-                    IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_INTERVAL_SETTING_KEY,
-                    TimeValue.timeValueMillis(cacheCleanIntervalInMillis)
-                )
-        );
-        Client client = client(node);
-        String index1 = "index1";
-        String index2 = "index2";
-        setupIndex(client, index1);
-        setupIndex(client, index2);
-
-        // create first cache entry in index1
-        createCacheEntry(client, index1, "hello");
-        assertCacheState(client, index1, 0, 1);
-        long memorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(memorySizeForIndex1 > 0);
-
-        // create second cache entry in index1
-        createCacheEntry(client, index1, "there");
-        assertCacheState(client, index1, 0, 2);
-        long finalMemorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(finalMemorySizeForIndex1 > memorySizeForIndex1);
-
-        // create first cache entry in index2
-        createCacheEntry(client, index2, "hello");
-        assertCacheState(client, index2, 0, 1);
-        assertTrue(getRequestCacheStats(client, index2).getMemorySizeInBytes() > 0);
-
-        // force refresh so that it creates 1 stale key
-        flushAndRefresh(index2);
-        // sleep until cache cleaner would have cleaned up the stale key from index 2
-        assertBusy(() -> {
-            // cache cleaner should have cleaned up the stale key from index 2
-            assertEquals(0, getRequestCacheStats(client, index2).getMemorySizeInBytes());
-            // cache cleaner should NOT have cleaned from index 1
-            assertEquals(finalMemorySizeForIndex1, getRequestCacheStats(client, index1).getMemorySizeInBytes());
-        }, cacheCleanIntervalInMillis * 2, TimeUnit.MILLISECONDS);
-    }
-
-    // when staleness threshold is higher than staleness, it should NOT clean the cache
-    public void testCacheCleanupSkipsWithHighStalenessThreshold() throws Exception {
-        int cacheCleanIntervalInMillis = 1;
-        String node = internalCluster().startNode(
-            Settings.builder()
-                .put(IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_STALENESS_THRESHOLD_SETTING_KEY, 0.90)
-                .put(
-                    IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_INTERVAL_SETTING_KEY,
-                    TimeValue.timeValueMillis(cacheCleanIntervalInMillis)
-                )
-        );
-        Client client = client(node);
-        String index1 = "index1";
-        String index2 = "index2";
-        setupIndex(client, index1);
-        setupIndex(client, index2);
-
-        // create first cache entry in index1
-        createCacheEntry(client, index1, "hello");
-        assertCacheState(client, index1, 0, 1);
-        long memorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(memorySizeForIndex1 > 0);
-
-        // create second cache entry in index1
-        createCacheEntry(client, index1, "there");
-        assertCacheState(client, index1, 0, 2);
-        long finalMemorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(finalMemorySizeForIndex1 > memorySizeForIndex1);
-
-        // create first cache entry in index2
-        createCacheEntry(client, index2, "hello");
-        assertCacheState(client, index2, 0, 1);
-        assertTrue(getRequestCacheStats(client, index2).getMemorySizeInBytes() > 0);
-
-        // force refresh so that it creates 1 stale key
-        flushAndRefresh(index2);
-        // sleep until cache cleaner would have cleaned up the stale key from index 2
-        assertBusy(() -> {
-            // cache cleaner should NOT have cleaned up the stale key from index 2
-            assertTrue(getRequestCacheStats(client, index2).getMemorySizeInBytes() > 0);
-            // cache cleaner should NOT have cleaned from index 1
-            assertEquals(finalMemorySizeForIndex1, getRequestCacheStats(client, index1).getMemorySizeInBytes());
-        }, cacheCleanIntervalInMillis * 2, TimeUnit.MILLISECONDS);
-    }
-
-    // when staleness threshold is explicitly set to 0, cache cleaner regularly cleans up stale keys.
-    public void testCacheCleanupOnZeroStalenessThreshold() throws Exception {
-        int cacheCleanIntervalInMillis = 50;
-        String node = internalCluster().startNode(
-            Settings.builder()
-                .put(IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_STALENESS_THRESHOLD_SETTING_KEY, 0)
-                .put(
-                    IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_INTERVAL_SETTING_KEY,
-                    TimeValue.timeValueMillis(cacheCleanIntervalInMillis)
-                )
-        );
-        Client client = client(node);
-        String index1 = "index1";
-        String index2 = "index2";
-        setupIndex(client, index1);
-        setupIndex(client, index2);
-
-        // create 10 index1 cache entries
-        for (int i = 1; i <= 10; i++) {
-            long cacheSizeBefore = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-            createCacheEntry(client, index1, "hello" + i);
-            assertCacheState(client, index1, 0, i);
-            long cacheSizeAfter = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-            assertTrue(cacheSizeAfter > cacheSizeBefore);
-        }
-
-        long finalMemorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-
-        // create first cache entry in index2
-        createCacheEntry(client, index2, "hello");
-        assertCacheState(client, index2, 0, 1);
-        assertTrue(getRequestCacheStats(client, index2).getMemorySizeInBytes() > 0);
-
-        // force refresh so that it creates 1 stale key
-        flushAndRefresh(index2);
-        // sleep until cache cleaner would have cleaned up the stale key from index 2
-        assertBusy(() -> {
-            // cache cleaner should have cleaned up the stale key from index 2
-            assertEquals(0, getRequestCacheStats(client, index2).getMemorySizeInBytes());
-            // cache cleaner should NOT have cleaned from index 1
-            assertEquals(finalMemorySizeForIndex1, getRequestCacheStats(client, index1).getMemorySizeInBytes());
-        }, cacheCleanIntervalInMillis * 2, TimeUnit.MILLISECONDS);
-    }
-
-    // when staleness threshold is not explicitly set, cache cleaner regularly cleans up stale keys
-    public void testStaleKeysRemovalWithoutExplicitThreshold() throws Exception {
-        int cacheCleanIntervalInMillis = 1;
-        String node = internalCluster().startNode(
-            Settings.builder()
-                .put(
-                    IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_INTERVAL_SETTING_KEY,
-                    TimeValue.timeValueMillis(cacheCleanIntervalInMillis)
-                )
-        );
-        String index1 = "index1";
-        String index2 = "index2";
-        Client client = client(node);
-        setupIndex(client, index1);
-        setupIndex(client, index2);
-
-        // create first cache entry in index1
-        createCacheEntry(client, index1, "hello");
-        assertCacheState(client, index1, 0, 1);
-        long memorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(memorySizeForIndex1 > 0);
-
-        // create second cache entry in index1
-        createCacheEntry(client, index1, "there");
-        assertCacheState(client, index1, 0, 2);
-        long finalMemorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(finalMemorySizeForIndex1 > memorySizeForIndex1);
-
-        // create first cache entry in index2
-        createCacheEntry(client, index2, "hello");
-        assertCacheState(client, index2, 0, 1);
-        assertTrue(getRequestCacheStats(client, index2).getMemorySizeInBytes() > 0);
-
-        // force refresh so that it creates 1 stale key
-        flushAndRefresh(index2);
-        // sleep until cache cleaner would have cleaned up the stale key from index 2
-        assertBusy(() -> {
-            // cache cleaner should have cleaned up the stale key from index 2
-            assertEquals(0, getRequestCacheStats(client, index2).getMemorySizeInBytes());
-            // cache cleaner should NOT have cleaned from index 1
-            assertEquals(finalMemorySizeForIndex1, getRequestCacheStats(client, index1).getMemorySizeInBytes());
-        }, cacheCleanIntervalInMillis * 2, TimeUnit.MILLISECONDS);
-    }
-
-    // when cache cleaner interval setting is not set, cache cleaner is configured appropriately with the fall-back setting
-    public void testCacheCleanupWithDefaultSettings() throws Exception {
-        int cacheCleanIntervalInMillis = 1;
-        String node = internalCluster().startNode(
-            Settings.builder().put(INDICES_CACHE_CLEANUP_INTERVAL_SETTING_KEY, TimeValue.timeValueMillis(cacheCleanIntervalInMillis))
-        );
-        Client client = client(node);
-        String index1 = "index1";
-        String index2 = "index2";
-        setupIndex(client, index1);
-        setupIndex(client, index2);
-
-        // create first cache entry in index1
-        createCacheEntry(client, index1, "hello");
-        assertCacheState(client, index1, 0, 1);
-        long memorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(memorySizeForIndex1 > 0);
-
-        // create second cache entry in index1
-        createCacheEntry(client, index1, "there");
-        assertCacheState(client, index1, 0, 2);
-        long finalMemorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(finalMemorySizeForIndex1 > memorySizeForIndex1);
-
-        // create first cache entry in index2
-        createCacheEntry(client, index2, "hello");
-        assertCacheState(client, index2, 0, 1);
-        assertTrue(getRequestCacheStats(client, index2).getMemorySizeInBytes() > 0);
-
-        // force refresh so that it creates 1 stale key
-        flushAndRefresh(index2);
-        // sleep until cache cleaner would have cleaned up the stale key from index 2
-        assertBusy(() -> {
-            // cache cleaner should have cleaned up the stale key from index 2
-            assertEquals(0, getRequestCacheStats(client, index2).getMemorySizeInBytes());
-            // cache cleaner should NOT have cleaned from index 1
-            assertEquals(finalMemorySizeForIndex1, getRequestCacheStats(client, index1).getMemorySizeInBytes());
-        }, cacheCleanIntervalInMillis * 2, TimeUnit.MILLISECONDS);
-    }
-
-    // staleness threshold updates flows through to the cache cleaner
-    public void testDynamicStalenessThresholdUpdate() throws Exception {
-        int cacheCleanIntervalInMillis = 1;
-        String node = internalCluster().startNode(
-            Settings.builder()
-                .put(IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_STALENESS_THRESHOLD_SETTING_KEY, 0.90)
-                .put(
-                    IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_INTERVAL_SETTING_KEY,
-                    TimeValue.timeValueMillis(cacheCleanIntervalInMillis)
-                )
-        );
-        Client client = client(node);
-        String index1 = "index1";
-        String index2 = "index2";
-        setupIndex(client, index1);
-        setupIndex(client, index2);
-
-        // create first cache entry in index1
-        createCacheEntry(client, index1, "hello");
-        assertCacheState(client, index1, 0, 1);
-        long memorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(memorySizeForIndex1 > 0);
-
-        // create second cache entry in index1
-        createCacheEntry(client, index1, "there");
-        assertCacheState(client, index1, 0, 2);
-        assertTrue(getRequestCacheStats(client, index1).getMemorySizeInBytes() > memorySizeForIndex1);
-
-        // create first cache entry in index2
-        createCacheEntry(client, index2, "hello");
-        assertCacheState(client, index2, 0, 1);
-        long finalMemorySizeForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(finalMemorySizeForIndex1 > 0);
-
-        // force refresh so that it creates 1 stale key
-        flushAndRefresh(index2);
-        assertBusy(() -> {
-            // cache cleaner should NOT have cleaned up the stale key from index 2
-            assertTrue(getRequestCacheStats(client, index2).getMemorySizeInBytes() > 0);
-        }, cacheCleanIntervalInMillis * 2, TimeUnit.MILLISECONDS);
-
-        // Update indices.requests.cache.cleanup.staleness_threshold to "10%"
-        ClusterUpdateSettingsRequest updateSettingsRequest = new ClusterUpdateSettingsRequest();
-        updateSettingsRequest.persistentSettings(Settings.builder().put(INDICES_REQUEST_CACHE_STALENESS_THRESHOLD_SETTING.getKey(), 0.10));
-        assertAcked(client().admin().cluster().updateSettings(updateSettingsRequest).actionGet());
-
-        assertBusy(() -> {
-            // cache cleaner should have cleaned up the stale key from index 2
-            assertEquals(0, getRequestCacheStats(client, index2).getMemorySizeInBytes());
-            // cache cleaner should NOT have cleaned from index 1
-            assertEquals(finalMemorySizeForIndex1, getRequestCacheStats(client, index1).getMemorySizeInBytes());
-        }, cacheCleanIntervalInMillis * 2, TimeUnit.MILLISECONDS);
-    }
-
-    // staleness threshold dynamic updates should throw exceptions on invalid input
-    public void testInvalidStalenessThresholdUpdateThrowsException() throws Exception {
-        // Update indices.requests.cache.cleanup.staleness_threshold to "10%" with illegal argument
-        assertThrows("Ratio should be in [0-1.0]", IllegalArgumentException.class, () -> {
-            ClusterUpdateSettingsRequest updateSettingsRequest = new ClusterUpdateSettingsRequest();
-            updateSettingsRequest.persistentSettings(
-                Settings.builder().put(IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_STALENESS_THRESHOLD_SETTING_KEY, 10)
-            );
-            client().admin().cluster().updateSettings(updateSettingsRequest).actionGet();
-        });
-    }
-
-    // closing the Index after caching will clean up from Indices Request Cache
-    public void testCacheClearanceAfterIndexClosure() throws Exception {
-        int cacheCleanIntervalInMillis = 100;
-        String node = internalCluster().startNode(
-            Settings.builder()
-                .put(IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_STALENESS_THRESHOLD_SETTING_KEY, 0.10)
-                .put(
-                    IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_INTERVAL_SETTING_KEY,
-                    TimeValue.timeValueMillis(cacheCleanIntervalInMillis)
-                )
-        );
-        Client client = client(node);
-        String index = "index";
-        setupIndex(client, index);
-
-        // assert there are no entries in the cache for index
-        assertEquals(0, getRequestCacheStats(client, index).getMemorySizeInBytes());
-        // assert there are no entries in the cache from other indices in the node
-        assertEquals(0, getNodeCacheStats(client).getMemorySizeInBytes());
-        // create first cache entry in index
-        createCacheEntry(client, index, "hello");
-        assertCacheState(client, index, 0, 1);
-        assertTrue(getRequestCacheStats(client, index).getMemorySizeInBytes() > 0);
-        assertTrue(getNodeCacheStats(client).getMemorySizeInBytes() > 0);
-
-        // close index
-        assertAcked(client.admin().indices().prepareClose(index));
-        // request cache stats cannot be access since Index should be closed
-        try {
-            getRequestCacheStats(client, index);
-        } catch (Exception e) {
-            assert (e instanceof IndexClosedException);
-        }
-        // sleep until cache cleaner would have cleaned up the stale key from index
-        assertBusy(() -> {
-            // cache cleaner should have cleaned up the stale keys from index
-            assertEquals(0, getNodeCacheStats(client).getMemorySizeInBytes());
-        }, cacheCleanIntervalInMillis * 2, TimeUnit.MILLISECONDS);
-    }
-
-    // deleting the Index after caching will clean up from Indices Request Cache
-    public void testCacheCleanupAfterIndexDeletion() throws Exception {
-        int cacheCleanIntervalInMillis = 100;
-        String node = internalCluster().startNode(
-            Settings.builder()
-                .put(IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_STALENESS_THRESHOLD_SETTING_KEY, 0.10)
-                .put(
-                    IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_INTERVAL_SETTING_KEY,
-                    TimeValue.timeValueMillis(cacheCleanIntervalInMillis)
-                )
-        );
-        Client client = client(node);
-        String index = "index";
-        setupIndex(client, index);
-
-        // assert there are no entries in the cache for index
-        assertEquals(0, getRequestCacheStats(client, index).getMemorySizeInBytes());
-        // assert there are no entries in the cache from other indices in the node
-        assertEquals(0, getNodeCacheStats(client).getMemorySizeInBytes());
-        // create first cache entry in index
-        createCacheEntry(client, index, "hello");
-        assertCacheState(client, index, 0, 1);
-        assertTrue(getRequestCacheStats(client, index).getMemorySizeInBytes() > 0);
-        assertTrue(getNodeCacheStats(client).getMemorySizeInBytes() > 0);
-
-        // delete index
-        assertAcked(client.admin().indices().prepareDelete(index));
-        // request cache stats cannot be access since Index should be deleted
-        try {
-            getRequestCacheStats(client, index);
-        } catch (Exception e) {
-            assert (e instanceof IndexNotFoundException);
-        }
-
-        // sleep until cache cleaner would have cleaned up the stale key from index
-        assertBusy(() -> {
-            // cache cleaner should have cleaned up the stale keys from index
-            assertEquals(0, getNodeCacheStats(client).getMemorySizeInBytes());
-        }, cacheCleanIntervalInMillis * 2, TimeUnit.MILLISECONDS);
-    }
-
-    // when staleness threshold is lower than staleness, it should clean cache from all indices having stale keys
-    public void testStaleKeysCleanupWithMultipleIndices() throws Exception {
-        int cacheCleanIntervalInMillis = 10;
-        String node = internalCluster().startNode(
-            Settings.builder()
-                .put(IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_STALENESS_THRESHOLD_SETTING_KEY, 0.10)
-                .put(
-                    IndicesRequestCache.INDICES_REQUEST_CACHE_CLEANUP_INTERVAL_SETTING_KEY,
-                    TimeValue.timeValueMillis(cacheCleanIntervalInMillis)
-                )
-        );
-        Client client = client(node);
-        String index1 = "index1";
-        String index2 = "index2";
-        setupIndex(client, index1);
-        setupIndex(client, index2);
-
-        // assert cache is empty for index1
-        assertEquals(0, getRequestCacheStats(client, index1).getMemorySizeInBytes());
-        // create first cache entry in index1
-        createCacheEntry(client, index1, "hello");
-        assertCacheState(client, index1, 0, 1);
-        long memorySizeForIndex1With1Entries = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(memorySizeForIndex1With1Entries > 0);
-
-        // create second cache entry in index1
-        createCacheEntry(client, index1, "there");
-        assertCacheState(client, index1, 0, 2);
-        long memorySizeForIndex1With2Entries = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-        assertTrue(memorySizeForIndex1With2Entries > memorySizeForIndex1With1Entries);
-
-        // assert cache is empty for index2
-        assertEquals(0, getRequestCacheStats(client, index2).getMemorySizeInBytes());
-        // create first cache entry in index2
-        createCacheEntry(client, index2, "hello");
-        assertCacheState(client, index2, 0, 1);
-        assertTrue(getRequestCacheStats(client, index2).getMemorySizeInBytes() > 0);
-
-        // force refresh both index1 and index2
-        flushAndRefresh(index1, index2);
-        // create another cache entry in index 1 same as memorySizeForIndex1With1Entries, this should not be cleaned up.
-        createCacheEntry(client, index1, "hello");
-        // sleep until cache cleaner would have cleaned up the stale key from index2
-        assertBusy(() -> {
-            // cache cleaner should have cleaned up the stale key from index2 and hence cache should be empty
-            assertEquals(0, getRequestCacheStats(client, index2).getMemorySizeInBytes());
-            // cache cleaner should have only cleaned up the stale entities for index1
-            long currentMemorySizeInBytesForIndex1 = getRequestCacheStats(client, index1).getMemorySizeInBytes();
-            // assert the memory size of index1 to only contain 1 entry added after flushAndRefresh
-            assertEquals(memorySizeForIndex1With1Entries, currentMemorySizeInBytesForIndex1);
-            // cache for index1 should not be empty since there was an item cached after flushAndRefresh
-            assertTrue(currentMemorySizeInBytesForIndex1 > 0);
-        }, cacheCleanIntervalInMillis * 2, TimeUnit.MILLISECONDS);
-    }
-
     public void testDeleteAndCreateSameIndexShardOnSameNode() throws Exception {
         String node_1 = internalCluster().startNode(Settings.builder().build());
         Client client = client(node_1);
@@ -1237,7 +688,12 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
         logger.info("Creating an index: {} with 2 shards", indexName);
         createIndex(
             indexName,
-            Settings.builder().put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 2).put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0).build()
+            Settings.builder()
+                .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 2)
+                .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
+                // Disable index refreshing to avoid cache being invalidated mid-test
+                .put(IndexSettings.INDEX_REFRESH_INTERVAL_SETTING.getKey(), TimeValue.timeValueMillis(-1))
+                .build()
         );
 
         ensureGreen(indexName);
@@ -1245,6 +701,9 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
         logger.info("Writing few docs and searching those which will cache items in RequestCache");
         indexRandom(true, client.prepareIndex(indexName).setSource("k", "hello"));
         indexRandom(true, client.prepareIndex(indexName).setSource("y", "hello again"));
+        ensureSearchable(indexName);
+        // Force merge the index to ensure there can be no background merges during the subsequent searches that would invalidate the cache
+        forceMerge(client, indexName);
         SearchResponse resp = client.prepareSearch(indexName).setRequestCache(true).setQuery(QueryBuilders.termQuery("k", "hello")).get();
         assertSearchResponse(resp);
         resp = client.prepareSearch(indexName).setRequestCache(true).setQuery(QueryBuilders.termQuery("y", "hello")).get();
@@ -1317,29 +776,10 @@ public class IndicesRequestCacheIT extends ParameterizedStaticSettingsOpenSearch
         return paths;
     }
 
-    private void setupIndex(Client client, String index) throws Exception {
-        assertAcked(
-            client.admin()
-                .indices()
-                .prepareCreate(index)
-                .setMapping("k", "type=keyword")
-                .setSettings(
-                    Settings.builder()
-                        .put(IndicesRequestCache.INDEX_CACHE_REQUEST_ENABLED_SETTING.getKey(), true)
-                        .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 1)
-                        .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0)
-                )
-                .get()
-        );
-        indexRandom(true, client.prepareIndex(index).setSource("k", "hello"));
-        indexRandom(true, client.prepareIndex(index).setSource("k", "there"));
-        ensureSearchable(index);
-    }
-
-    private void createCacheEntry(Client client, String index, String value) {
-        SearchResponse resp = client.prepareSearch(index).setRequestCache(true).setQuery(QueryBuilders.termQuery("k", value)).get();
-        assertSearchResponse(resp);
-        OpenSearchAssertions.assertAllSuccessful(resp);
+    private void forceMerge(Client client, String index) {
+        ForceMergeResponse forceMergeResponse = client.admin().indices().prepareForceMerge(index).setFlush(true).get();
+        OpenSearchAssertions.assertAllSuccessful(forceMergeResponse);
+        refreshAndWaitForReplication();
     }
 
     private static void assertCacheState(Client client, String index, long expectedHits, long expectedMisses) {
